Use new version of lastpass-java
The new version of lastpass-java uses google-http-client instead of
jersey-client, which means the implicit dependencies on the JAXB apis
etc. have been removed.

Also removed the flags from dex which were keeping services from META-INF
and marking this .apk as a "Core Library" to allow us to bundle the javax
classes, which are no longer required.

A side effect of this is that the build time and resulting .apk size have
both decreased.